+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 
+### Fixed
+- Usage of deprecated classes
+
 ## [0.3.2] - 2018-12-10
 ### Added
 - Add support for Symfony 4