+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Mark all classes as `final`
 - Add return type hints
+- Print an empty string with twig `breakpoint` function
 
 ### Removed
 - Drop support for PHP<7.4