+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 - Drop support for PHP<7.4
 - Drop support for Symfony<4.4
+- Drop support for Twig<2.15.3
 
 ## [0.3.5] - 2021-02-08